fix: biobank-scale hygiene — OOM guards, fp16 max_post, dead code removal
- em.py warm-start path (chroms 2+): gate jnp.array(geno) with
  fits_on_device, matching the chrom-1 fix from de8ac2a
- hmm.py: allocate max_post as float16 in both forward_backward_decode
  and forward_backward_bucketed_decode (was float32 — 108 GB at T=25k)
- em.py: delete init_model_from_labels (uncalled, would OOM via
  geno.astype(float32) at biobank scale)
- demo.py: gate geno device upload with fits_on_device
- output.py: fail-fast in write_global_ancestry when posteriors fallback
  is triggered at H > 100k (would allocate H*T*A*4 bytes)
- docs/BIOBANK_SCALE.md: add 9-item memory traps section, update fixed table

+++ popout/hmm.py
@@ -1276,7 +1276,7 @@ def forward_backward_decode(
     emit_pad = pc["emit_pad"]
 
     calls = _np.zeros((H, T), dtype=_np.int8)
-    max_post = _np.zeros((H, T), dtype=_np.float32) if compute_max_post else None
+    max_post = _np.zeros((H, T), dtype=_np.float16) if compute_max_post else None
     global_sums = _np.zeros((H, A), dtype=_np.float64) if compute_max_post else None
 
     for start in range(0, H, batch_size):
@@ -1323,7 +1323,7 @@ def forward_backward_bucketed_decode(
     B = len(model.bucket_centers)
 
     calls = _np.zeros((H, T), dtype=_np.int8)
-    max_post = _np.zeros((H, T), dtype=_np.float32) if compute_max_post else None
+    max_post = _np.zeros((H, T), dtype=_np.float16) if compute_max_post else None
     global_sums = _np.zeros((H, A), dtype=_np.float64) if compute_max_post else None
 
     bucket_np = _np.array(model.bucket_assignments)
